+++ packages/cloud-cognitive/src/components/ProductiveCard/ProductiveCard.js
@@ -9,11 +9,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Card } from '../Card';
 import { pkg } from '../../settings';
-import { stripUnwantedProps } from '../../global/js/utils/props-helper';
+import { prepareProps } from '../../global/js/utils/props-helper';
 const componentName = 'ProductiveCard';
 
 export let ProductiveCard = (props) => {
-  const validProps = stripUnwantedProps(props, [
+  const validProps = prepareProps(props, [
     'media',
     'mediaPosition',
     'onSecondaryButtonClick',
@@ -36,7 +36,10 @@ ProductiveCard.propTypes = {
   actionIcons: PropTypes.arrayOf(
     PropTypes.shape({
       id: PropTypes.string,
-      icon: PropTypes.object,
+      icon: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
+      onKeyDown: PropTypes.func,
+      onClick: PropTypes.func,
+      iconDescription: PropTypes.string,
     })
   ),
   /**
@@ -79,6 +82,7 @@ ProductiveCard.propTypes = {
       id: PropTypes.string,
       itemText: PropTypes.string,
       onClick: PropTypes.func,
+      onKeyDown: PropTypes.func,
     })
   ),
   /**